Factor out TraceGesture component.
